Generalize the timezone select component
